requirements.txt
Django==6.0
psycopg[binary]
Pillow>=11.0.0
django-crispy-forms==2.3
crispy-bootstrap5==2025.6
django-widget-tweaks==1.5.0
django-import-export==4.0.1
openpyxl==3.1.5
pandas>=2.2.3
python-docx==1.1.0
django-filter==24.3
django-extensions==3.2.3
django-environ==0.11.2
whitenoise==6.7.0
gunicorn>=23.0.0
python-decouple==3.8
coverage==7.4.0
pytest-django==4.7.0
factory-boy==3.3.0